kay.ai/requirements.txt
aiohttp==3.10.5
anthropic==0.49.0
fastapi==0.115.12
openai==1.69.0
psycopg2-binary==2.9.10
pydantic==2.11.0
python-dotenv==1.1.0
Requests==2.32.3
weaviate_client==4.11.1
fastapi[standard]==0.115.12